pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aquaguard-rl"
version = "1.0.0"
description = "India Groundwater & Agricultural Resource Management RL Environment — Meta PyTorch OpenEnv Hackathon"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ashunaukari01@gmail.com", email = "ashunaukari01@gmail.com"}
]
keywords = [
    "reinforcement-learning",
    "openenv",
    "groundwater",
    "agriculture",
    "india",
    "climate",
    "policy",
    "llm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "aquaguard_rl.server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
hf = [
    "huggingface-hub>=0.23.0",
]

[project.urls]
Repository = "https://github.com/Ashgen12/AquaGuard-RL"
"Hugging Face Space" = "https://huggingface.co/spaces/Ashgen12/AquaGuard-RL"
"OpenEnv" = "https://github.com/meta-pytorch/OpenEnv"

[tool.setuptools]
include-package-data = true
packages = ["aquaguard_rl", "aquaguard_rl.server"]
package-dir = { "aquaguard_rl" = ".", "aquaguard_rl.server" = "server" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 100
target-version = "py311"